Add files via upload
a) Creating an index on movie_id and sorts the keys in the index in ascending
order. Verify the query plan
b) Show various indexes created on movie collection.
c) Sort movie_id in descending order.
d) Create a descending order index on movie_id to get ratings related to “Toy
Story (1995)” verify the query plan.
e) Limit the number of items in the result of above query.